+++ lib/improver/tests/utilities/cube_manipulation/test_enforce_coordinate_ordering.py
@@ -38,7 +38,8 @@ from iris.cube import Cube
 from iris.exceptions import CoordinateNotFoundError
 from iris.tests import IrisTest
 
-from improver.utilities.cube_manipulation import enforce_coordinate_ordering
+from improver.utilities.cube_manipulation_new import (
+    enforce_coordinate_ordering)
 
 from improver.tests.ensemble_calibration.ensemble_calibration.\
     helper_functions import set_up_temperature_cube
