+++ lib/improver/tests/utilities/cube_manipulation/test_get_filtered_attributes.py
@@ -37,7 +37,7 @@ import numpy as np
 
 from iris.tests import IrisTest
 
-from improver.utilities.cube_manipulation import get_filtered_attributes
+from improver.utilities.cube_manipulation_new import get_filtered_attributes
 from improver.tests.set_up_test_cubes import set_up_variable_cube
