+++ lib/improver/tests/utilities/cube_manipulation/test_merge_cubes.py
@@ -33,23 +33,16 @@ Unit tests for the function "cube_manipulation.merge_cubes".
 """
 
 import unittest
+import numpy as np
 from datetime import datetime as dt
 
 import iris
 from iris.cube import Cube
 from iris.exceptions import DuplicateDataError, MergeError
 from iris.tests import IrisTest
-import numpy as np
 
 from improver.utilities.cube_checker import find_threshold_coordinate
-from improver.utilities.cube_manipulation import merge_cubes
-
-from improver.tests.ensemble_calibration.ensemble_calibration.\
-    helper_functions import (
-        set_up_temperature_cube,
-        set_up_probability_above_threshold_temperature_cube,
-        add_forecast_reference_time_and_forecast_period)
-
+from improver.utilities.cube_manipulation_new import merge_cubes
 from improver.utilities.warnings_handler import ManageWarnings
 from improver.tests.set_up_test_cubes import (
     set_up_variable_cube, set_up_probability_cube)
@@ -61,38 +54,35 @@ class Test_merge_cubes(IrisTest):
 
     def setUp(self):
         """Use temperature cube to test with."""
-        self.cube = set_up_temperature_cube()
-        self.cube_ukv = self.cube.extract(iris.Constraint(realization=1))
-        self.cube_ukv.remove_coord('realization')
-        self.cube_ukv.attributes['mosg__grid_type'] = 'standard'
-        self.cube_ukv.attributes['mosg__model_configuration'] = 'uk_det'
-        self.cube_ukv.attributes['mosg__grid_domain'] = 'uk_extended'
-        self.cube_ukv.attributes['mosg__grid_version'] = '1.2.0'
-        self.cube_ukv_t1 = self.cube_ukv.copy()
-        self.cube_ukv_t2 = self.cube_ukv.copy()
-        add_forecast_reference_time_and_forecast_period(self.cube_ukv,
-                                                        fp_point=4.0)
-        add_forecast_reference_time_and_forecast_period(self.cube_ukv_t1,
-                                                        fp_point=5.0)
-        add_forecast_reference_time_and_forecast_period(self.cube_ukv_t2,
-                                                        fp_point=6.0)
-        add_forecast_reference_time_and_forecast_period(self.cube,
-                                                        fp_point=7.0)
-        self.cube.attributes['mosg__grid_type'] = 'standard'
-        self.cube.attributes['mosg__model_configuration'] = 'uk_ens'
-        self.cube.attributes['mosg__grid_domain'] = 'uk_extended'
-        self.cube.attributes['mosg__grid_version'] = '1.2.0'
-        self.prob_ukv = set_up_probability_above_threshold_temperature_cube()
-        self.prob_ukv.attributes['mosg__grid_type'] = 'standard'
-        self.prob_ukv.attributes['mosg__model_configuration'] = 'uk_det'
-        self.prob_ukv.attributes['mosg__grid_domain'] = 'uk_extended'
-        self.prob_ukv.attributes['mosg__grid_version'] = '1.2.0'
-        self.prob_enuk = set_up_probability_above_threshold_temperature_cube()
-        self.prob_enuk.attributes.update({'mosg__grid_type': 'standard'})
-        self.prob_enuk.attributes.update(
-            {'mosg__model_configuration': 'uk_ens'})
-        self.prob_enuk.attributes.update({'mosg__grid_domain': 'uk_extended'})
-        self.prob_enuk.attributes.update({'mosg__grid_version': '1.2.0'})
+
+        data = np.array([[[226.15, 237.4, 248.65],
+                          [259.9, 271.15, 282.4],
+                          [293.65, 304.9, 316.15]],
+                         [[230.15, 241.4, 252.65],
+                          [263.9, 275.15, 286.4],
+                          [297.65, 308.9, 320.15]],
+                         [[232.15, 243.4, 254.65],
+                          [265.9, 277.15, 288.4],
+                          [299.65, 310.9, 322.15]]], dtype=np.float32)
+
+        # set up a MOGREPS-UK cube with 7 hour forecast period
+        time_point = dt(2015, 11, 23, 7)
+        self.cube = set_up_variable_cube(
+            data.copy(), standard_grid_metadata='uk_ens', time=time_point,
+            frt=dt(2015, 11, 23, 0))
+
+        # set up a UKV cube with 4 hour forecast period
+        self.cube_ukv = set_up_variable_cube(
+            data[1].copy(), standard_grid_metadata='uk_det', time=time_point,
+            frt=dt(2015, 11, 23, 3))
+
+        # set up more UKV cubes with 5 and 6 hour forecast periods
+        self.cube_ukv_t1 = set_up_variable_cube(
+            data[1].copy(), standard_grid_metadata='uk_det', time=time_point,
+            frt=dt(2015, 11, 23, 2))
+        self.cube_ukv_t2 = set_up_variable_cube(
+            data[1].copy(), standard_grid_metadata='uk_det', time=time_point,
+            frt=dt(2015, 11, 23, 1))
 
         # Setup two non-Met Office model example configuration cubes.
         # Using a simple temperature data array, one cube set is setup
@@ -102,22 +92,15 @@ class Test_merge_cubes(IrisTest):
         self.data_3d = np.array([self.data, self.data, self.data])
 
         self.cube_non_mo_det = set_up_variable_cube(self.data)
-
         self.cube_non_mo_ens = set_up_variable_cube(
             self.data_3d, realizations=np.array([0, 3, 4]))
 
         self.cube_non_mo_det.attributes['non_mo_model_config'] = 'non_uk_det'
         self.cube_non_mo_ens.attributes['non_mo_model_config'] = 'non_uk_ens'
 
-    @ManageWarnings(record=True)
-    def test_basic(self, warning_list=None):
+    def test_basic(self):
         """Test that the utility returns an iris.cube.Cube."""
-        result = merge_cubes(self.cube)
-        self.assertTrue(any(item.category == UserWarning
-                            for item in warning_list))
-        warning_msg = "Only a single cube "
-        self.assertTrue(any(warning_msg in str(item)
-                            for item in warning_list))
+        result = merge_cubes([self.cube_ukv, self.cube_ukv_t1])
         self.assertIsInstance(result, Cube)
 
     def test_identical_cubes(self):
@@ -128,127 +111,15 @@ class Test_merge_cubes(IrisTest):
             merge_cubes(cubes)
 
     def test_lagged_ukv(self):
-        """Test Lagged ukv merge OK"""
+        """Test lagged UKV merge OK (forecast periods in seconds)"""
+        expected_fp_points = 3600*np.array([6, 5, 4], dtype=np.int32)
         cubes = iris.cube.CubeList([self.cube_ukv,
                                     self.cube_ukv_t1,
                                     self.cube_ukv_t2])
         result = merge_cubes(cubes)
         self.assertIsInstance(result, Cube)
         self.assertArrayAlmostEqual(
-            result.coord("forecast_period").points, [6.0, 5.0, 4.0])
-
-    def test_multi_model(self):
-        """Test Multi models merge OK"""
-        cubes = iris.cube.CubeList([self.cube, self.cube_ukv])
-        result = merge_cubes(cubes, model_id_attr="mosg__model_configuration")
-        self.assertIsInstance(result, Cube)
-        self.assertArrayAlmostEqual(
-            result.coord("model_realization").points, [0., 1.,
-                                                       2., 1000.])
-
-    def test_no_model_id_attr_multi_model(self):
-        """Test multiple model blending fails and results in a merge error if
-        no model_id_attr is specified."""
-        cubes = iris.cube.CubeList([self.cube, self.cube_ukv])
-        with self.assertRaises(MergeError):
-            merge_cubes(cubes)
-
-    def test_threshold_data(self):
-        """Test threshold data merges OK"""
-        cubes = iris.cube.CubeList([self.prob_ukv, self.prob_enuk])
-        result = merge_cubes(cubes, model_id_attr="mosg__model_configuration")
-        self.assertArrayAlmostEqual(
-            result.coord("model_id").points, [0., 1000.])
-
-    def test_non_mo_model_id(self):
-        """Test that a model ID attribute string can be specified when
-        merging multi model cubes"""
-        cubes = iris.cube.CubeList(
-            [self.cube_non_mo_ens, self.cube_non_mo_det])
-        result = merge_cubes(cubes, model_id_attr='non_mo_model_config')
-        self.assertIsInstance(result, Cube)
-        self.assertArrayAlmostEqual(
-            result.coord(
-                "model_realization").points, [0., 3., 4., 1000.])
-
-    def test_model_id_attr_mismatch(self):
-        """Test that when a model ID attribute string is specified that does
-        not match the model ID attribute key name on both cubes to be merged,
-        an error is thrown"""
-        cubes = iris.cube.CubeList(
-            [self.cube_non_mo_ens, self.cube_non_mo_det])
-
-        # The test cubes contain the 'non_mo_model_config' attribute key.
-        # We'll specify 'non_matching_model_config' as our model ID
-        # attribute key argument. Merge_cubes should then raise a warning
-        # as our specified model ID does not match that on the cubes and it
-        # will not be able to build a model ID coordinate.
-        msg = ('Cannot create model ID coordinate for grid blending '
-               'as the model ID attribute specified is not found '
-               'within the cube attributes')
-
-        with self.assertRaisesRegex(ValueError, msg):
-            merge_cubes(cubes, model_id_attr='non_matching_model_config')
-
-    def test_model_id_attr_mismatch_one_cube(self):
-        """Test that when a model ID attribute string is specified that only
-        matches the model ID attribute key name on one of the cubes to be
-        merged, an error is thrown"""
-
-        # Change the model ID attribute key on one of the test cubes so that
-        # it matches the model ID argument. Merge_cubes should still raise
-        # an error as the model ID attribute key has to match on all cubes
-        # to be blended.
-        self.cube_non_mo_det.attributes.pop('non_mo_model_config')
-        self.cube_non_mo_det.attributes[
-            'non_matching_model_config'] = 'non_uk_det'
-
-        cubes = iris.cube.CubeList(
-            [self.cube_non_mo_ens, self.cube_non_mo_det])
-
-        msg = ('Cannot create model ID coordinate for grid blending '
-               'as the model ID attribute specified is not found '
-               'within the cube attributes')
-
-        with self.assertRaisesRegex(ValueError, msg):
-            merge_cubes(cubes, model_id_attr='non_matching_model_config')
-
-    def test_one_threshold_data(self):
-        """Test threshold data where one cube has single threshold as dim"""
-        ukv_prob = self.prob_ukv[0]
-        threshold_coord = find_threshold_coordinate(ukv_prob).name()
-        ukv_prob = iris.util.new_axis(ukv_prob, threshold_coord)
-        enuk_prob = self.prob_enuk[0]
-        cubes = iris.cube.CubeList([ukv_prob, enuk_prob])
-        result = merge_cubes(cubes, model_id_attr='mosg__model_configuration')
-        self.assertArrayAlmostEqual(
-            result.coord("model_id").points, [0., 1000.])
-        self.assertEqual(ukv_prob.data.shape, (1, 1, 3, 3))
-        self.assertEqual(enuk_prob.data.shape, (1, 3, 3))
-        self.assertEqual(result.data.shape, (2, 3, 3))
-
-    def test_mismatched_time_bounds_ranges(self):
-        """Test for mismatched bounds ranges error."""
-        frt = dt(2017, 11, 9, 21, 0)
-        times = [dt(2017, 11, 10, 3, 0),
-                 dt(2017, 11, 10, 4, 0),
-                 dt(2017, 11, 10, 5, 0)]
-        time_bounds = np.array([
-            [dt(2017, 11, 10, 2, 0), dt(2017, 11, 10, 3, 0)],
-            [dt(2017, 11, 10, 3, 0), dt(2017, 11, 10, 4, 0)],
-            [dt(2017, 11, 10, 2, 0), dt(2017, 11, 10, 5, 0)]])
-
-        cubelist = iris.cube.CubeList([])
-        for tpoint, tbounds in zip(times, time_bounds):
-            cube = set_up_probability_cube(
-                0.6*np.ones((2, 3, 3), dtype=np.float32),
-                np.array([278., 280.], dtype=np.float32),
-                time=tpoint, frt=frt, time_bounds=tbounds)
-            cubelist.append(cube)
-
-        msg = "Cube with mismatching time bounds ranges"
-        with self.assertRaisesRegex(ValueError, msg):
-            merge_cubes(cubelist, blend_coord="time")
+            result.coord("forecast_period").points, expected_fp_points)
 
 
 if __name__ == '__main__':
