--- lib/improver/tests/utilities/cube_manipulation_new/test_strip_var_names.py
@@ -1,95 +1,0 @@
-# -*- coding: utf-8 -*-
-# -----------------------------------------------------------------------------
-# (C) British Crown Copyright 2017-2019 Met Office.
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-# * Redistributions of source code must retain the above copyright notice, this
-#   list of conditions and the following disclaimer.
-#
-# * Redistributions in binary form must reproduce the above copyright notice,
-#   this list of conditions and the following disclaimer in the documentation
-#   and/or other materials provided with the distribution.
-#
-# * Neither the name of the copyright holder nor the names of its
-#   contributors may be used to endorse or promote products derived from
-#   this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-"""
-Unit tests for the function "cube_manipulation.strip_var_names".
-"""
-
-import unittest
-
-import iris
-from iris.tests import IrisTest
-
-from improver.utilities.cube_manipulation_new import strip_var_names
-
-from improver.tests.ensemble_calibration.ensemble_calibration.\
-    helper_functions import set_up_temperature_cube
-
-
-class Test_strip_var_names(IrisTest):
-
-    """Test the _slice_var_names utility."""
-
-    def setUp(self):
-        """Use temperature cube to test with."""
-        self.cube = set_up_temperature_cube()
-
-    def test_basic(self):
-        """Test that the utility returns an iris.cube.CubeList."""
-        self.cube.var_name = "air_temperature"
-        result = strip_var_names(self.cube)
-        self.assertIsInstance(result, iris.cube.CubeList)
-
-    def test_cube_var_name_is_none(self):
-        """
-        Test that the utility returns an iris.cube.Cube with a
-        var_name of None.
-        """
-        self.cube.var_name = "air_temperature"
-        result = strip_var_names(self.cube)
-        self.assertIsNone(result[0].var_name, None)
-
-    def test_cube_coord_var_name_is_none(self):
-        """
-        Test that the coordinates have var_names of None.
-        """
-        self.cube.coord("time").var_name = "time"
-        self.cube.coord("latitude").var_name = "latitude"
-        self.cube.coord("longitude").var_name = "longitude"
-        result = strip_var_names(self.cube)
-        for cube in result:
-            for coord in cube.coords():
-                self.assertIsNone(coord.var_name, None)
-
-    def test_cubelist(self):
-        """Test that the utility returns an iris.cube.CubeList."""
-        cube1 = self.cube
-        cube2 = self.cube
-        cubes = iris.cube.CubeList([cube1, cube2])
-        self.cube.var_name = "air_temperature"
-        result = strip_var_names(cubes)
-        self.assertIsInstance(result, iris.cube.CubeList)
-        for cube in result:
-            for coord in cube.coords():
-                self.assertIsNone(coord.var_name, None)
-
-
-if __name__ == '__main__':
-    unittest.main()
